Fix the tokenize() pickle issue
