+++ train_classifier.py
@@ -6,14 +6,8 @@ import numpy as np
 import sqlite3
 from sqlalchemy import create_engine
 import os
-import re
 import pickle
-import nltk
-nltk.download(['punkt','stopwords','wordnet'])
-from nltk.corpus import stopwords
-from nltk.tokenize import word_tokenize
-from nltk.stem.porter import PorterStemmer
-from nltk.stem.wordnet import WordNetLemmatizer
+from tokenizer_util import tokenize
 
 from sklearn.pipeline import Pipeline,FeatureUnion
 from sklearn.feature_extraction.text import CountVectorizer, TfidfTransformer
@@ -37,25 +31,6 @@ def load_data(database_filepath):
     Y = df.drop(['id', 'message', 'original', 'genre'], axis = 1)
     return X,Y,category_names
 
-def tokenize(text):
-    ''' Natural Language Processing: Normalize, Tokenize, Stem/Lemmatize
-    - Normalize (lowercase, punctuation)
-    - Tokenize words
-    - Remove stop words (a, the ...)
-    - Lemmatize
-    '''
-    # Convert text to lowercase and remove punctuation
-    text = text.lower()
-    text = re.sub(r"[^a-zA-Z0-9]", " ", text)    
-    # Tokenize words
-    tokens = word_tokenize(text)    
-    # Remove Stop words, Stem & Lemmed words
-    stop_word = stopwords.words("english")
-    stemmer = PorterStemmer()
-    lemmatizer = WordNetLemmatizer()    
-    #stemmed = [stemmer.stem(w) for w in tokens if w not in stop_word]
-    lemmed = [lemmatizer.lemmatize(w) for w in tokens if w not in stop_word]
-    return lemmed 
 
 def build_model():
     ''' Model:
@@ -73,6 +48,7 @@ def build_model():
         'tfidf__smooth_idf': [True, False]
     }
     model  = GridSearchCV(pipeline, param_grid = parameters, cv=2) 
+    model = pipeline
     return model  
 
 def evaluate_model(model, X_test, Y_test, category_names):
